requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
flask==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pytest==8.4.1
python-decouple==3.8
requests==2.32.4
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
werkzeug==3.1.3
